+++ src/lib/.babelrc.js
@@ -17,7 +17,7 @@ const config = {
 
 if (process.env.PREPARE_MODULE_TYPE)
   config.ignore = [
-    `./src/index.${process.env.PREPARE_MODULE_TYPE === 'esm' ? 'js' : 'mjs'}`,
+    `./index.${process.env.PREPARE_MODULE_TYPE === 'esm' ? 'js' : 'mjs'}`,
   ];
 
 module.exports = config;